BaseTest.cs --------------------------- Added function that makes it easier for tests to create connection strings with and without database
StoredProcedure.cs
------------------------------
Bug #11450  	Connector/Net, current database and stored procedures

Fixed this bug by checking for the presence of dot (.) in the spname.  If it is there, then the part before the dot is assumed to be the database name.  If not dot is present, then the function database() is used to make sure we execute the procedure or function from the current database.

Testsute/storedprocedure.cs
---------------------------------------------
Added testcase for Bug #11450  	Connector/Net, current database and stored procedures