fixed Bug #8382  	Commandbuilder does not handle queries to other databases than the default one-
command.cs
Now checking to see if result.AffectedRows != -1 when deciding if we should update our update count.  We need to do this
because sometimes we have non-selects and selects in the same batch and the result.IsResultSet will return true when we 
actually do have an update count.

commandbuilder.cs
-----------------------------
cleaned up our generateschema method.  

commandresult.cs
----------------------------
moved settting of readSchema and readRows to false inside the while loop.  We did this because the datareader close method calls back to the commandresult
consume method a second time.  The way the code was would have the those variables set to false and therefore the consume method would try
and read off another resultset, hanging the system.

Also, updating the internal affectedRows count much more cleanly.  Now it's -1 if we are all selects and only goes > -1 if there is a non-select.

datareader.cs
-------------------------
Setting baseschemaname to the database name now

Driver.cs
-------------------
In configure, we are making sure our server props hash actually has keys for character_set_client and character_set_connection before we refereence them
changed the sig of ReadResult so that rows is long so we can return -1 for select result

nativedriver.cs
-------------------------
ReadResult properly returns affectedRows as -1 for selects and > -1 for non-selects

parameter.cs
--------------------------
Parameter no longer infers type once you set the type. As long as you never use one of the type setting properties, then setting the value will continue to infer the type.

CommandBuilderTests.cs
-----------------------------------------
Added test case for bug #8382