Bug #19261  	Supplying Input Parameters
This bug was caused by the PreparedStatement class using the wrong parameter array for setting the null map.  It was running over the parameter array given by the user (which can be out of order) howeve the null map has the same order as the parameters returned in the metadata from the PREPARE call.  So the fix was to use the metadata parameter array to get the parameter name, and then index into the user given parameter array to determine if the parameter is null.