Bug #18570  	Unsigned tinyint (NET byte) incorrectly determined param type from param val
This bug was causes by two things.  (1) not correctly setting the mysqldbtype to ubyte and (2) sending the wrong data type to the server for prepared statements

parameter.cs - Changed GetPSType() to send the code for Byte or'ed with 0x8000 (to make it unsigned).  Also, fixed SetDbTYpe code to properly set the type to UByte for unsigned tinyint values.
PreparedStatements.cs - Added test case for this bug