Bug #19481 Where clause with datetime throws exception [any warning causes the exception]
The problem here is that the driver must issue a sql query to the server to retrieve the warnings and this uses a data reader.  Only one reader can be open at one time so if you had a query that might return more than one resultset, then at
the end of the first it tried to retrieve the warnings.  This failed since there is already a reader open.  The solution
was to wait until the reader is closed to report the warnings.  This is suboptimal since the warning callbacks will not occur until you close but it is the only easy way to solve it.  The other way would be to retrieve the warnings without
using a data reader but that would be alot more effort.