Bug #15077 Error MySqlCommandBuilder.DeriveParameters for sp without parameters.
The error was a simple usage of a null reference.  When there are no parameters for a stored proc,
then GetParameterList returns null but the following code doesn't check that.  It assumes that what
is returned is an array of zero or more elements.  The fix is to only process the array if it is non-null.