Bug #17814 Stored procedure fails unless DbType set explicitly
The problem here is that our current implementation of output variables relies on user variables that don't preserve type.
The current code did have discovery capabililty but that was only being used for the DeriveParameters method.
The fix here is to call DiscoverParameters on a copy of the command only in the case that the user did not 
fully specify the types of all the parameters.  This is suboptimal and it is suggested that the user either call
DeriveParameters on their own or fully specify the types to prevent the additional work.