Bug #23268 System.FormatException when invoking procedure with ENUM input parameter
This was fixed by adding enum to the list of types that is checked in the parameter parsing code.  The old code was assuming the parameter was decimal which led to the format exception.  As a result of this fix, the procedure parameters collection returned by GetSchema now has each parameter name including the prevailing parameter marker.  This matches what SQLClient does.