Bug #25912 selecting negative time values gets wrong results
The problem here is that we are using a TimeSpan struct to hold time values.  The ctor for TimeSpan applies each value to the total time value.  So passing in -7 hours along with a positive 24 minutes results in -7 hours + 24 minutes which is -6 hours and 36 minutes.  The fix was to negate each value if the hours value is negative.