Bug #25603 Critial ConnectionPool Error in Connector.Net 5.03
Fixed problem where pool was not checking the connection for being alive unless connection reset was given on the connection string.  It should always check that the connection is alive.  Also, fixed a problem in GetConnection that resulted from our recent overhaul of pooling.  It was only making a single pass looking for a connetion.