Bug #25569 UpdateRowSource.FirstReturnedRecord does not work
Fixed this by exposing a property on the command builder where the user can request that generated keys be returned and then applying a final select later in the process.  Thanks go to Lynn Eriksen for some good feedback on this bug and for submitting an idea for how it could be fixed.