Bug #26347	Update commands not being generated with TableAdapter wizard
The problem was caused by the confusion between schema and catalog.  We use schema to mean database and indicate that we do not support catalog.  However, in our MySqlDataSourceInformation class we were not returning null for the DefaultCatalog property. This was causing the QPA inside Visual studio to return a table name of the form <schema>.<schema>.<table_name> which obviously is not going to work.  We also needed to fix a problem in parameter name generation in our commandbuilder class.