- fixed problem that caused in use connection strings to be modified when a pooled connection timed out and was cancelled.  (bug #40091)
The problem was that we were creating our connection object before setting the pooling property to false.  This caused a connection to be made in an existing connection string cache which caused all the connection strings in that cache to change when we set our pooling property.  The fix was to create a connection string builder object, set the pooling property, and then create the connection object.